requirements_updated.txt
# ==========================================
# Core Framework (FastAPI + Async)
# ==========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
Flask==2.3.3
flask-cors==4.0.0

# ==========================================
# ML & ML - Phase 2 Enhancement
# ==========================================
google-generativeai==0.3.0
transformers==4.45.2
torch==2.5.0
sentence-transformers==2.2.2
scikit-learn==1.3.2

# ==========================================
# Multimodal Analysis
# ==========================================
Pillow==10.1.0
opencv-python==4.8.1.78
python-magic-bin==0.4.14
PyPDF2==3.0.1

# ==========================================
# Database
# ==========================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# ==========================================
# Caching
# ==========================================
redis==5.0.1
aioredis==2.0.1

# ==========================================
# Monitoring & Observability
# ==========================================
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-jaeger==1.21.0
python-json-logger==2.0.7

# ==========================================
# Production Server
# ==========================================
gunicorn==21.2.0

# ==========================================
# Development
# ==========================================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.12.1
flake8==6.1.0
mypy==1.7.1
